+++ neurom/core/population.py
@@ -28,6 +28,7 @@
 
 """Morphology Population Classes and Functions."""
 import logging
+import os
 from pathlib import Path
 
 from morphio import MorphioError
@@ -40,17 +41,7 @@ L = logging.getLogger(__name__)
 
 def _resolve_if_morphology_paths(files_or_objects):
     """Resolve the files in the list."""
-
-    def resolve_path(path):
-        """Resolve only the parent of the file path.
-
-        Not resolving the filename allows to have symlinks to files with different filename, but
-        any symlink in the parent tree would be correctly resolved.
-        """
-        path = Path(path)
-        return path.parent.resolve() / path.name
-
-    return [resolve_path(f) if isinstance(f, (Path, str)) else f for f in files_or_objects]
+    return [Path(os.path.abspath(f)) if isinstance(f, (Path, str)) else f for f in files_or_objects]
 
 
 class Population:
@@ -74,6 +65,9 @@ class Population:
                 will be loaded everytime it is accessed within the population. Which is good when
                 population is big. If true then all morphs will be loaded upon the construction
                 and kept in memory.
+
+        Notes:
+            symlinks in paths are not resolved.
         """
         self._ignored_exceptions = ignored_exceptions
         self.name = name